% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_exposure_data}
\alias{add_exposure_data}
\title{Combine titre data and infection data into biokinetics input format.}
\usage{
add_exposure_data(dat_sero, dat_inf, exposure_column = "day")
}
\arguments{
\item{dat_sero}{data.table containing titre values in the format required by biokinetics. See data vignette: \code{vignette("data", package = "epikinetics")}.}

\item{dat_inf}{data.table containing exposure days and person ids corresponding to those in dat_sero. By default the exposure days are expected in a column called 'day'.}

\item{exposure_column}{Default 'day'. The name of the column containing exposure days. These can be integers or dates.}
}
\value{
A data.table with all columns required by the biokinetics model.
}
\description{
The biokinetics class requires a single data table with titre readings
and last exposure times for all individuals. If you have exposure times and titre readings for the same
set of individuals in separate files, this function will combine them into a single data.table.
}
