% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{biokinetics_priors}
\alias{biokinetics_priors}
\title{Construct priors for the biomarker model.}
\usage{
biokinetics_priors(
  mu_values = c(4, 10, 60, 0.25, -0.02, 0),
  sigma_values = c(2, 2, 3, 0.01, 0.01, 0.01)
)
}
\arguments{
\item{mu_values}{Mean of Gaussian prior for each of t0, tp, ts, m1, m2, m3, in order.}

\item{sigma_values}{Standard deviation of Gaussian prior for each of t0, tp, ts, m1, m2, m3, in order.}
}
\value{
A named list of type 'biokinetics_priors'.
}
\description{
The biokinetics model has 6 parameters: t0, tp, ts, m1, m2, m3 corresponding to critical time points and
gradients. See the model vignette for details: \code{vignette("model", package = "epikinetics")}. Each of these
parameters has a Gaussian prior, and these can be specified by the user. This function takes means and standard
deviations for each prior and constructs an object of type 'biokinetics_priors' to be passed to the model.
}
\examples{
priors <- biokinetics_priors(mu_values = c(4.0, 10, 60, 0.25, -0.02, 0),
sigma_values = c(2.0, 2.0, 3.0, 0.01, 0.01, 0.01))
}
