% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{cmdstan_fit}
\alias{cmdstan_fit}
\title{Access the underlying CmdStanR fit}
\usage{
cmdstan_fit(x)
}
\arguments{
\item{x}{An \code{epikinetics_fit} object.}
}
\value{
The underlying \code{CmdStanMCMC} object.
}
\description{
Access the underlying CmdStanR fit
}
