% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_log2_scale}
\alias{convert_log2_scale}
\title{Base 2 log scale conversion}
\usage{
convert_log2_scale(dt_in, smallest_value, vars_to_transform = "value")
}
\arguments{
\item{dt_in}{data.table containing data to be transformed to base 2 log scale.}

\item{smallest_value}{The lowest value in the original dataset. Used to get the data onto
a scale that starts at zero.}

\item{vars_to_transform}{Names of columns to apply the transformation to.}
}
\value{
A data.table, identical to the input data but with specified columns transformed.
}
\description{
Natural scale data is converted to a base 2 log scale before model fitting.
This function does not modify the provided data.table in-place, but returns a transformed copy.
}
