% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_log_scale_inverse}
\alias{convert_log_scale_inverse}
\title{Invert base 2 log scale conversion}
\usage{
convert_log_scale_inverse(dt_in, vars_to_transform)
}
\arguments{
\item{dt_in}{data.table containing data to be transformed from base 2 log to natural scale.}

\item{vars_to_transform}{Names of columns to apply the transformation to.}
}
\value{
A data.table, identical to the input data but with specified columns transformed.
}
\description{
User provided data is converted to a base 2 log scale before model fitting. This
function reverses that transformation. This function does not modify the provided data.table in-place,
but returns a transformed copy.
}
