% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{gaussian_priors}
\alias{gaussian_priors}
\title{Construct Gaussian priors for model variables.}
\usage{
gaussian_priors(names, mu_values, sigma_values)
}
\arguments{
\item{names}{Names of the variables.}

\item{mu_values}{Mean of Gaussian prior for each named variable.}

\item{sigma_values}{Standard deviation of Gaussian prior for each named variable.}
}
\value{
A named list of type 'gaussian_priors'.
}
\description{
Construct an object of type 'gaussian_priors' to pass to the model.
}
\examples{
priors <- gaussian_priors(names = c("t0", "tp"), mu_values = c(4.0, 10), sigma_values = c(2.0, 2.0))
}
