% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.biokinetics_individual_trajectories}
\alias{plot.biokinetics_individual_trajectories}
\title{Plot method for "biokinetics_individual_trajectories" class}
\usage{
\method{plot}{biokinetics_individual_trajectories}(
  x,
  ...,
  data = NULL,
  min_day = NULL,
  max_day = NULL,
  pids = NULL,
  titre_types = NULL
)
}
\arguments{
\item{x}{An object of class "biokinetics_individual_trajectories". These are
generated by running biokinetics$simulate_individual_trajectories(). See
\href{../../epikinetics/html/biokinetics.html#method-biokinetics-simulate_individaul_trajectories}{\code{biokinetics$simulate_individual_trajectories()}}}

\item{\dots}{Further arguments passed to the method.}

\item{min_day}{Optional minimum date}

\item{max_day}{Optional maximum date}

\item{titre_types}{Optional vector of titre types to include.}

\item{pid}{Optional vector of ids to plot simulated trajectories for a subset of individuals. Can only be used
if x has been generated with summarise=FALSE.}
}
\description{
Plot method for "biokinetics_individual_trajectories" class
}
