% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.biokinetics_population_stationary_points}
\alias{plot.biokinetics_population_stationary_points}
\title{Plot method for "biokinetics_population_stationary_points" class}
\usage{
\method{plot}{biokinetics_population_stationary_points}(x, ..., upper_detection_limit = NULL)
}
\arguments{
\item{x}{An object of class "biokinetics_population_stationary_points". These are
generated by running biokinetics$population_stationary_points(). See
\href{../../epikinetics/html/biokinetics.html#method-biokinetics-population_stationary_points}{\code{biokinetics$population_stationary_points()}}}

\item{\dots}{Further arguments passed to the method.}

\item{upper_detection_limit}{Numeric. Optional upper detection limit. Will be plotted as a dotted line.}
}
\description{
Plot method for "biokinetics_population_stationary_points" class
}
