% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.biokinetics_population_trajectories}
\alias{plot.biokinetics_population_trajectories}
\title{Plot method for "biokinetics_population_trajectories" class}
\usage{
\method{plot}{biokinetics_population_trajectories}(x, ..., data = NULL)
}
\arguments{
\item{x}{An object of class "biokinetics_population_trajectories". These are
generated by running biokinetics$simulate_populate_trajectories(). See
\href{../../epikinetics/html/biokinetics.html#method-biokinetics-simulate_population_trajectories}{\code{biokinetics$simulate_population_trajectories()}}}

\item{\dots}{Further arguments passed to the method.}

\item{data}{Optional data.table containing raw data as provided to the biokinetics model.}
}
\description{
Plot method for "biokinetics_population_trajectories" class
}
