% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.biokinetics_priors}
\alias{plot.biokinetics_priors}
\title{Simulate biomarker kinetics predicted by the given biokinetics priors
and optionally compare to a dataset.}
\usage{
\method{plot}{biokinetics_priors}(
  x,
  ...,
  tmax = 150,
  n_draws = 2000,
  data = NULL,
  upper_detection_limit = NULL,
  lower_detection_limit = NULL
)
}
\arguments{
\item{x}{A named list of type 'biokinetics_priors'.}

\item{\dots}{Further arguments passed to the method.}

\item{tmax}{Integer. The number of time points in each simulated trajectory. Default 150.}

\item{n_draws}{Integer. The number of trajectories to simulate. Default 2000.}

\item{data}{Optional data.frame with columns time_since_last_exp and value. The raw data to compare to.}

\item{upper_detection_limit}{Optional upper detection limit.}

\item{lower_detection_limit}{Optional lower detection limit.}
}
\value{
A ggplot2 object.
}
\description{
Simulate trajectories by drawing random samples from the given
priors for each parameter in the biokinetics model.
}
