% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prior_predictive}
\alias{plot_prior_predictive}
\title{Simulate biomarker kinetics predicted by the given biokinetics priors
and optionally compare to a dataset.}
\usage{
plot_prior_predictive(priors, tmax = 150, n_draws = 2000, data = NULL)
}
\arguments{
\item{priors}{A named list of type 'biokinetics_priors'.}

\item{tmax}{Integer. The number of time points in each simulated trajectory. Default 150.}

\item{n_draws}{Integer. The number of trajectories to simulate. Default 2000.}

\item{data}{Optional data.frame with columns t_since_last_exp and value. The raw data to compare to.}
}
\value{
A ggplot2 object.
}
\description{
Simulate trajectories by drawing random samples from the given
priors for each parameter in the biokinetics model.
}
