% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prepare_stan_data}
\alias{prepare_stan_data}
\title{Prepare data for stan model.}
\usage{
prepare_stan_data(data, priors, covariate_formula, preds_sd, time_type)
}
\arguments{
\item{data}{Data table of model inputs.}

\item{priors}{Object of type 'epikinetic_priors'.}

\item{covariate_formula}{Formula specifying hierarchical structure of model.}

\item{time_type}{One of 'relative' or 'absolute'.}
}
\value{
A named list.
}
\description{
Shape the data to named list format for stan.
}
