% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{process_fits}
\alias{process_fits}
\title{Process model fits.}
\usage{
process_fits(
  fits,
  data,
  covariate_formula = ~0 + infection_history,
  time_type = "relative",
  t_max = 150,
  summarise = TRUE,
  by = c("infection_history", "titre_type"),
  scale = "natural",
  n_draws = 2500
)
}
\arguments{
\item{fits}{A CmdStanMCMC fitted model object.}

\item{data}{Data table. The original data used to fit the model.}

\item{covariate_formula}{Formula specifying hierarchical structure of model. Ddefault  ~0 + infection_history.}

\item{time_type}{One of 'relative' or 'absolute'. Default 'relative'.}

\item{t_max}{Numeric}

\item{summarise}{Boolean}

\item{by}{Vector of covariates to summarise by. Can be any of 'titre_type' plus covariates specified
in covariate_formula. Default c("infection_history", "titre_type").}

\item{scale}{One of 'natural' or 'log'. Default 'natural'.}

\item{n_draws}{Integer. Number of samples to draw. Default 2500.}
}
\value{
A Table.
}
\description{
Process the stan model results into a data table.
}
