% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_model}
\alias{run_model}
\title{Fit the model.}
\usage{
run_model(
  priors,
  data = NULL,
  file_path = NULL,
  covariate_formula = ~0 + infection_history,
  preds_sd = 0.25,
  time_type = "relative",
  ...
)
}
\arguments{
\item{priors}{Object of type 'epikinetic_priors'.}

\item{data}{Optional data table of model inputs. One of data or file must be provided.}

\item{file_path}{Optional file path to model inputs in CSV format. One of data or file must be provided.}

\item{covariate_formula}{Formula specifying hierarchical structure of model. Default ~0 + infection_history.}

\item{preds_sd}{Standard deviation of predictor coefficients. Default 0.25.}

\item{time_type}{One of 'relative' or 'absolute'. Default 'relative'.}

\item{...}{Named arguments to the \code{sample()} method of CmdStan model.
objects: \url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}}
}
\value{
A CmdStanMCMC fitted model object.
}
\description{
Fit the kinetics model and return fitted model.
}
