% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\title{SARS-CoV-2 Antibody Kinetics Model}
\description{
Create an instance of the SARS-CoV-2 antibody model and
fit it to a dataset using Gaussian priors and a given hierarchical model structure.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sam-new}{\code{sam$new()}}
\item \href{#method-sam-fit}{\code{sam$fit()}}
\item \href{#method-sam-population_trajectories}{\code{sam$population_trajectories()}}
\item \href{#method-sam-population_stationary_points}{\code{sam$population_stationary_points()}}
\item \href{#method-sam-individual_trajectories}{\code{sam$individual_trajectories()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sam-new"></a>}}
\if{latex}{\out{\hypertarget{method-sam-new}{}}}
\subsection{Method \code{new()}}{
Initialise the kinetics model and return fitted model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sam$new(
  priors = sam_priors(),
  data = NULL,
  file_path = NULL,
  covariate_formula = ~0 + infection_history,
  preds_sd = 0.25,
  time_type = "relative"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{priors}}{Object of type 'sam_priors'. Default sam_priors().}

\item{\code{data}}{Optional data table of model inputs. One of data or file must be provided.}

\item{\code{file_path}}{Optional file path to model inputs in CSV format. One of data or file must be provided.}

\item{\code{covariate_formula}}{Formula specifying hierarchical structure of model. Default ~0 + infection_history.}

\item{\code{preds_sd}}{Standard deviation of predictor coefficients. Default 0.25.}

\item{\code{time_type}}{One of 'relative' or 'absolute'. Default 'relative'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An epikinetics::sam object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sam-fit"></a>}}
\if{latex}{\out{\hypertarget{method-sam-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model and return CmdStanMCMC fitted model object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sam$fit(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments to the \code{sample()} method of CmdStan model.
objects: \url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A CmdStanMCMC fitted model object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sam-population_trajectories"></a>}}
\if{latex}{\out{\hypertarget{method-sam-population_trajectories}{}}}
\subsection{Method \code{population_trajectories()}}{
Process the model results into a data table of titre values over time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sam$population_trajectories(
  time_type = "relative",
  t_max = 150,
  summarise = TRUE,
  scale = "natural",
  n_draws = 2500
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_type}}{One of 'relative' or 'absolute'. Default 'relative'.}

\item{\code{t_max}}{Integer. Maximum number of time points to include.}

\item{\code{summarise}}{Boolean. Default TRUE. If TRUE returns values for 0.025, 0.5 and 0.975 quantiles, if FALSE returns
individual values.}

\item{\code{scale}}{One of 'natural' or 'log'. Default 'natural'.}

\item{\code{n_draws}}{Integer. Number of samples to draw. Default 2500.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Table containing titre values at time points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sam-population_stationary_points"></a>}}
\if{latex}{\out{\hypertarget{method-sam-population_stationary_points}{}}}
\subsection{Method \code{population_stationary_points()}}{
Process the stan model results into a data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sam$population_stationary_points(n_draws = 2500)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_draws}}{Integer. Number of samples to draw. Default 2500.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sam-individual_trajectories"></a>}}
\if{latex}{\out{\hypertarget{method-sam-individual_trajectories}{}}}
\subsection{Method \code{individual_trajectories()}}{
Simulate individual trajectories from the model. This is
computationally expensive and may take a while to run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sam$individual_trajectories(
  n_draws,
  time_shift,
  scale = "log",
  adjust_dates = FALSE,
  add_variation_params = FALSE,
  t_max = 150
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_draws}}{Integer. Number of samples to draw.}

\item{\code{time_shift}}{Integer.}

\item{\code{scale}}{One of 'natural' or 'log'. Default 'log'.}

\item{\code{adjust_dates}}{Logical. Default FALSE.}

\item{\code{add_variation_params}}{Logical. Default FALSE.}

\item{\code{t_max}}{Integer. Maximum number of time points to include. Default 150.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Table.
}
}
}
