% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{sam_priors}
\alias{sam_priors}
\title{Construct priors for the SARS-CoV-2 antibody model.}
\usage{
sam_priors(
  mu_values = c(4, 10, 60, 0.25, -0.02, 0),
  sigma_values = c(2, 2, 3, 0.01, 0.01, 0.01)
)
}
\arguments{
\item{mu_values}{Mean of Gaussian prior for each of t0, tp, ts, m1, m2, m3, in order.}

\item{sigma_values}{Standard deviation of Gaussian prior for each of t0, tp, ts, m1, m2, m3, in order.}
}
\value{
A named list of type 'sam_priors'.
}
\description{
Construct an object of type 'sam_priors' to pass to the model.
}
\examples{
priors <- sam_priors(mu_values = c(4.0, 10, 60, 0.25, -0.02, 0), sigma_values = c(2.0, 2.0, 3.0, 0.01, 0.01, 0.01))
}
