% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biokinetics.R
\name{biokinetics}
\alias{biokinetics}
\title{Biomarker Kinetics Model}
\description{
Create an instance of the biomarker kinetics model and
fit it to a dataset.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-biokinetics-new}{\code{biokinetics$new()}}
\item \href{#method-biokinetics-get_stan_data}{\code{biokinetics$get_stan_data()}}
\item \href{#method-biokinetics-fit}{\code{biokinetics$fit()}}
\item \href{#method-biokinetics-extract_population_parameters}{\code{biokinetics$extract_population_parameters()}}
\item \href{#method-biokinetics-extract_individual_parameters}{\code{biokinetics$extract_individual_parameters()}}
\item \href{#method-biokinetics-simulate_population_trajectories}{\code{biokinetics$simulate_population_trajectories()}}
\item \href{#method-biokinetics-population_stationary_points}{\code{biokinetics$population_stationary_points()}}
\item \href{#method-biokinetics-simulate_individual_trajectories}{\code{biokinetics$simulate_individual_trajectories()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biokinetics-new"></a>}}
\if{latex}{\out{\hypertarget{method-biokinetics-new}{}}}
\subsection{Method \code{new()}}{
Initialise the kinetics model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biokinetics$new(
  priors = biokinetics_priors(),
  data = NULL,
  file_path = NULL,
  covariate_formula = ~0,
  preds_sd = 0.25,
  time_type = "relative"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{priors}}{Object of type \link[epikinetics]{biokinetics_priors}. Default biokinetics_priors().}

\item{\code{data}}{Optional data table of model inputs. One of data or file must be provided. See the data vignette
for required columns: \code{vignette("data", package = "epikinetics")}.}

\item{\code{file_path}}{Optional file path to model inputs in CSV format. One of data or file must be provided.}

\item{\code{covariate_formula}}{Formula specifying hierarchical structure of model. Default ~0.}

\item{\code{preds_sd}}{Standard deviation of predictor coefficients. Default 0.25.}

\item{\code{time_type}}{One of 'relative' or 'absolute'. Default 'relative'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An epikinetics::biokinetics object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biokinetics-get_stan_data"></a>}}
\if{latex}{\out{\hypertarget{method-biokinetics-get_stan_data}{}}}
\subsection{Method \code{get_stan_data()}}{
View the data that is passed to the stan model, for debugging purposes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biokinetics$get_stan_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of arguments that will be passed to the stan model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biokinetics-fit"></a>}}
\if{latex}{\out{\hypertarget{method-biokinetics-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model and return CmdStanMCMC fitted model object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biokinetics$fit(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments to the \code{sample()} method of CmdStan model.
objects: \url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A CmdStanMCMC fitted model object: \url{https://mc-stan.org/cmdstanr/reference/CmdStanMCMC.html}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biokinetics-extract_population_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-biokinetics-extract_population_parameters}{}}}
\subsection{Method \code{extract_population_parameters()}}{
Extract fitted population parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biokinetics$extract_population_parameters(
  n_draws = 2500,
  human_readable_covariates = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_draws}}{Numeric}

\item{\code{human_readable_covariates}}{Logical. Default TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biokinetics-extract_individual_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-biokinetics-extract_individual_parameters}{}}}
\subsection{Method \code{extract_individual_parameters()}}{
Extract fitted individual parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biokinetics$extract_individual_parameters(
  n_draws = 2500,
  include_variation_params = TRUE,
  human_readable_covariates = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_draws}}{Numeric}

\item{\code{include_variation_params}}{Logical}

\item{\code{human_readable_covariates}}{Logical. Default TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biokinetics-simulate_population_trajectories"></a>}}
\if{latex}{\out{\hypertarget{method-biokinetics-simulate_population_trajectories}{}}}
\subsection{Method \code{simulate_population_trajectories()}}{
Process the model results into a data table of titre values over time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biokinetics$simulate_population_trajectories(
  time_type = "relative",
  t_max = 150,
  summarise = TRUE,
  n_draws = 2500
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_type}}{One of 'relative' or 'absolute'. Default 'relative'.}

\item{\code{t_max}}{Integer. Maximum number of time points to include.}

\item{\code{summarise}}{Boolean. Default TRUE. If TRUE, summarises over draws from posterior parameter distributions to
return 0.025, 0.5 and 0.975 quantiles, labelled lo, me and hi, respectively. If FALSE returns values for individual
draws from posterior parameter distributions.}

\item{\code{n_draws}}{Integer. Maximum number of samples to include. Default 2500.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table containing titre values at time points. If summarise = TRUE, columns are t, me, lo, hi,
titre_type, and a column for each covariate in the hierarchical model. If summarise = FALSE, columns are t, .draw
t0_pop, tp_pop, ts_pop, m1_pop, m2_pop, m3_pop, beta_t0, beta_tp, beta_ts, beta_m1, beta_m2, beta_m3, mu
titre_type and a column for each covariate in the hierarchical model. See the data vignette for details:
\code{vignette("data", package = "epikinetics")}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biokinetics-population_stationary_points"></a>}}
\if{latex}{\out{\hypertarget{method-biokinetics-population_stationary_points}{}}}
\subsection{Method \code{population_stationary_points()}}{
Process the stan model results into a data.table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biokinetics$population_stationary_points(n_draws = 2500)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_draws}}{Integer. Maximum number of samples to include. Default 2500.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table of peak and set titre values. Columns are tire_type, mu_p, mu_s, rel_drop_me, mu_p_me,
mu_s_me, and a column for each covariate. See the data vignette for details:
\code{vignette("data", package = "epikinetics")}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biokinetics-simulate_individual_trajectories"></a>}}
\if{latex}{\out{\hypertarget{method-biokinetics-simulate_individual_trajectories}{}}}
\subsection{Method \code{simulate_individual_trajectories()}}{
Simulate individual trajectories from the model. This is
computationally expensive and may take a while to run if n_draws is large.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biokinetics$simulate_individual_trajectories(
  summarise = TRUE,
  n_draws = 2500,
  time_shift = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summarise}}{Boolean. If TRUE, average the individual trajectories to get lo, me and
hi values for the population, disaggregated by titre type. If FALSE return the indidivudal trajectories.
Default TRUE.}

\item{\code{n_draws}}{Integer. Maximum number of samples to draw. Default 2500.}

\item{\code{time_shift}}{Integer. Number of days to adjust the exposure date by. Default 0.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table. If summarise = TRUE columns are calendar_date, titre_type, me, lo, hi, time_shift.
If summarise = FALSE, columns are pid, draw, t, mu, titre_type, exposure_date, calendar_date, time_shift
and a column for each covariate in the hierarchical model. See the data vignette for details:
\code{vignette("data", package = "epikinetics")}.
}
}
}
