% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epikinetics-package.R
\docType{package}
\name{epikinetics-package}
\alias{epikinetics}
\alias{epikinetics-package}
\title{epikinetics: Antibody and Cycle Threshold Kinetics Modelling}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alex Hill \email{alex.hill@gmail.com} [contributor]

Authors:
\itemize{
  \item Timothy Russell \email{timothy.russell@lshtm.ac.uk}
}

}
\keyword{internal}
