% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epikinetics-package.R
\docType{package}
\name{epikinetics-package}
\alias{epikinetics}
\alias{epikinetics-package}
\title{epikinetics: Biomarker Kinetics Modelling}
\description{
Fit kinetic curves to biomarker data, using a Bayesian hierarchical model.
}
\author{
\strong{Maintainer}: Alex Hill \email{alex.hill@gmail.com}

Authors:
\itemize{
  \item Timothy Russell \email{timothy.russell@lshtm.ac.uk}
}

}
\keyword{internal}
