% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{epikinetics_priors}
\alias{epikinetics_priors}
\title{Construct Gaussian priors for model variables.}
\usage{
epikinetics_priors(
  names = c("t0", "tp", "ts", "m1", "m2", "m3"),
  mu_values = c(4, 10, 60, 0.25, -0.02, 0),
  sigma_values = c(2, 2, 3, 0.01, 0.01, 0.01)
)
}
\arguments{
\item{names}{Names of the variables.}

\item{mu_values}{Mean of Gaussian prior for each named variable.}

\item{sigma_values}{Standard deviation of Gaussian prior for each named variable.}
}
\value{
A named list of type 'epikinetics_priors'.
}
\description{
Construct an object of type 'epikinetics_priors' to pass to the model.
}
\examples{
priors <- epikinetics_priors(names = c("t0", "tp", "ts", "m1", "m2", "m3"),
mu_values = c(4.0, 10, 60, 0.25, -0.02, 0), sigma_values = c(2.0, 2.0, 3.0, 0.01, 0.01, 0.01))
}
