% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sero_data}
\alias{plot_sero_data}
\title{Plot serological data}
\usage{
plot_sero_data(
  data,
  tmax = 150,
  covariates = character(0),
  upper_censoring_limit = NULL,
  lower_censoring_limit = NULL
)
}
\arguments{
\item{data}{A data.table with required columns time_since_last_exp, value and titre_type.}

\item{tmax}{Integer. The number of time points in each simulated trajectory. Default 150.}

\item{covariates}{Optional vector of covariate names to facet by (these must correspond to columns in the data.table)}

\item{upper_censoring_limit}{Optional upper detection limit.}

\item{lower_censoring_limit}{Optional lower detection limit.}
}
\value{
A ggplot2 object.
}
\description{
Plot serological data in the format provided to the biokinetics
model, with a smoothing function fitted.
}
