% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prepare_stan_data}
\alias{prepare_stan_data}
\title{Prepare data for stan model.}
\usage{
prepare_stan_data(dt, priors, formula, time_type, preds_sd)
}
\value{
A named list.
}
\description{
Shape the data to named list format for stan.
}
