% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{process_fits}
\alias{process_fits}
\title{Process model fits.}
\usage{
process_fits(
  fit,
  dt,
  covariate_formula = ~0 + infection_history,
  time_type = "relative",
  t_max = 150,
  summarise = TRUE,
  by = c("Infection history", "Titre type"),
  scale = "natural",
  cleaned_names = c("Infection history", "Titre type"),
  n_draws = 2500
)
}
\arguments{
\item{fit}{A CmdStanMCMC fitted model object.}

\item{dt}{Data table. The original data used to fit the model.}

\item{covariate_formula}{Formula specifying hierarchical structure of model. Ddefault  ~0 + infection_history.}

\item{time_type}{One of 'relative' or 'absolute'. Default 'relative'.}

\item{t_max}{Numeric}

\item{summarise}{Boolean}

\item{by}{Vector of covariates to summarise by. Values must match cleaned variable names.
Default c('Infection history', 'Titre type').}

\item{scale}{One of 'natural' or 'log'. Default 'natural'.}

\item{cleaned_names}{Vector of human readable variable names. Default c('Infection history', 'Titre type').}

\item{n_draws}{Integer. Number of samples to draw. Default 2500.}
}
\value{
A Table.
}
\description{
Process the stan model results into a data table.
}
