% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_model}
\alias{run_model}
\title{Fit the model.}
\usage{
run_model(
  data,
  priors,
  covariate_formula = ~0 + infection_history,
  preds_sd = 0.25,
  time_type = "relative",
  ...
)
}
\arguments{
\item{data}{Model inputs.}

\item{priors}{Object of type 'epikinetic_priors'.}

\item{time_type}{One of 'relative' or 'absolute'.}

\item{...}{Named arguments to the \code{sample()} method of CmdStan model
objects: \url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}}
}
\value{
A CmdStanMCMC fitted model object.
}
\description{
Fit the kinetics model and return fitted model.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  res <- run_model(data = dt, priors = epikinetic_priors(), chains = 4, parallel_chains = 4, iter_warmup = 100, iter_sampling = 400, threads_per_chain = 4)
}
}
