% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scova.R
\name{scova}
\alias{scova}
\title{SARS-CoV-2 Antibody Kinetics Model}
\description{
Create an instance of the SARS-CoV-2 antibody model and
fit it to a dataset using Gaussian priors and a given hierarchical model structure.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-scova-new}{\code{scova$new()}}
\item \href{#method-scova-fit}{\code{scova$fit()}}
\item \href{#method-scova-simulate_population_trajectories}{\code{scova$simulate_population_trajectories()}}
\item \href{#method-scova-population_stationary_points}{\code{scova$population_stationary_points()}}
\item \href{#method-scova-simulate_individual_trajectories}{\code{scova$simulate_individual_trajectories()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scova-new"></a>}}
\if{latex}{\out{\hypertarget{method-scova-new}{}}}
\subsection{Method \code{new()}}{
Initialise the kinetics model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scova$new(
  priors = scova_priors(),
  data = NULL,
  file_path = NULL,
  covariate_formula = ~0,
  preds_sd = 0.25,
  time_type = "relative"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{priors}}{Object of type 'scova_priors'. Default scova_priors().}

\item{\code{data}}{Optional data table of model inputs. One of data or file must be provided.}

\item{\code{file_path}}{Optional file path to model inputs in CSV format. One of data or file must be provided.}

\item{\code{covariate_formula}}{Formula specifying hierarchical structure of model. Default ~0.}

\item{\code{preds_sd}}{Standard deviation of predictor coefficients. Default 0.25.}

\item{\code{time_type}}{One of 'relative' or 'absolute'. Default 'relative'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An epikinetics::sam object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scova-fit"></a>}}
\if{latex}{\out{\hypertarget{method-scova-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model and return CmdStanMCMC fitted model object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scova$fit(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments to the \code{sample()} method of CmdStan model.
objects: \url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A CmdStanMCMC fitted model object: \url{https://mc-stan.org/cmdstanr/reference/CmdStanMCMC.html}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scova-simulate_population_trajectories"></a>}}
\if{latex}{\out{\hypertarget{method-scova-simulate_population_trajectories}{}}}
\subsection{Method \code{simulate_population_trajectories()}}{
Process the model results into a data table of titre values over time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scova$simulate_population_trajectories(
  time_type = "relative",
  t_max = 150,
  summarise = TRUE,
  n_draws = 2500
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_type}}{One of 'relative' or 'absolute'. Default 'relative'.}

\item{\code{t_max}}{Integer. Maximum number of time points to include.}

\item{\code{summarise}}{Boolean. Default TRUE. If TRUE returns values for 0.025, 0.5 and 0.975 quantiles, if FALSE returns
all values.}

\item{\code{n_draws}}{Integer. Maximum number of samples to include. Default 2500.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table containing titre values at time points. If summarise = TRUE, columns are t, p, k, me, lo, hi,
titre_type, and a column for each covariate in the hierarchical model. If summarise = FALSE, columns are t, p, k,
t0_pop, tp_pop, ts_pop, m1_pop, m2_pop, m3_pop, beta_t0, beta_tp, beta_ts, beta_m1, beta_m2, beta_m3, mu
.draw, titre_type and a column for each covariate in the hierarchical model.
lower and upper bounds of titre values; if summarise = FALSE then
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scova-population_stationary_points"></a>}}
\if{latex}{\out{\hypertarget{method-scova-population_stationary_points}{}}}
\subsection{Method \code{population_stationary_points()}}{
Process the stan model results into a data.table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scova$population_stationary_points(n_draws = 2500)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_draws}}{Integer. Maximum number of samples to include. Default 2500.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table of peak and set titre values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scova-simulate_individual_trajectories"></a>}}
\if{latex}{\out{\hypertarget{method-scova-simulate_individual_trajectories}{}}}
\subsection{Method \code{simulate_individual_trajectories()}}{
Simulate individual trajectories from the model. This is
computationally expensive and may take a while to run if n_draws is large.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scova$simulate_individual_trajectories(
  t_max = 150,
  summarise = TRUE,
  n_draws = 2500,
  time_shift = 0,
  add_variation_params = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_max}}{Integer. Maximum number of time points to include. Default 150.}

\item{\code{summarise}}{Boolean. If TRUE, average the individual trajectories to get lo, me and
hi values for the population. If FALSE return the simulated indidivudal trajectories.
Default TRUE.}

\item{\code{n_draws}}{Integer. Maximum number of samples to draw. Default 2500.}

\item{\code{time_shift}}{Integer. Number of days to adjust the exposure date by. Default 0.}

\item{\code{add_variation_params}}{Logical. Default FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table.
}
}
}
